Version 2.0
Version 2.0 upgrades the User Interface by attempting to strip out most
of the jquery mobile code and replacing it with standard Jquery.  It
also attempts to consolidate a lot of the separate pages into the
index.html file and displays each page in a hidden div which is
animated via rollup and rolldown to make a smoother user experience.